refactor: split shape model and geometry modules
