+++ labelme/_shape/__init__.py
@@ -1,0 +1,35 @@
+from .hit_testing import nearest_edge_index
+from .hit_testing import nearest_rotation_point_index
+from .hit_testing import nearest_vertex_index
+from .model import CIRCLE_POINT_COUNT
+from .model import LINE_POINT_COUNT
+from .model import MIN_LINESTRIP_POINT_COUNT
+from .model import MIN_POLYGON_POINT_COUNT
+from .model import ORIENTED_RECTANGLE_POINT_COUNT
+from .model import POLYLINE_SHAPE_TYPES
+from .model import RECTANGLE_POINT_COUNT
+from .model import Shape
+from .model import ShapeType
+from .oriented_rectangle import get_rotation_handle
+from .oriented_rectangle import oriented_rectangle_arrow_points
+from .oriented_rectangle import oriented_rectangle_center
+from .oriented_rectangle import rotate
+
+__all__ = [
+    "CIRCLE_POINT_COUNT",
+    "LINE_POINT_COUNT",
+    "MIN_LINESTRIP_POINT_COUNT",
+    "MIN_POLYGON_POINT_COUNT",
+    "ORIENTED_RECTANGLE_POINT_COUNT",
+    "POLYLINE_SHAPE_TYPES",
+    "RECTANGLE_POINT_COUNT",
+    "Shape",
+    "ShapeType",
+    "get_rotation_handle",
+    "nearest_edge_index",
+    "nearest_rotation_point_index",
+    "nearest_vertex_index",
+    "oriented_rectangle_arrow_points",
+    "oriented_rectangle_center",
+    "rotate",
+]
